feat(exportModal): add version field to exported flow data
The version field is added to the exported flow data in order to keep track of the version of the flow. This allows for better management and identification of different versions of the flow.

fix(exportModal): fetch and set version from API

The version is fetched from the API using the `getVersion` function and then set in the state variable `version` using the `setVersion` function. This ensures that the correct version is used when exporting the flow.

feat(flow): add version field to FlowType

The `FlowType` interface is updated to include a `version` field. This allows for storing and retrieving the version of the flow when needed.